requirements.txt
django
django-rest_framework
django-cors-headers
psycopg2
django-filter
openai
python-dotenv